feat(dsg):prepare context
