+++ packages/amplication-code-gen-types/src/plugin-events-params.ts
@@ -1,5 +1,6 @@
 import { namedTypes } from "ast-types";
 import { JsonValue } from "type-fest";
+import winston from "winston";
 import {
   Entity,
   EntityField,
@@ -7,7 +8,12 @@ import {
   PrismaDataSource,
 } from "./code-gen-types";
 import { EventParams } from "./plugins-types";
+import { DSGResourceData } from "./dsg-resource-data";
 
+export interface PrepareContextParams extends EventParams {
+  dSGResourceData: DSGResourceData;
+  logger: winston.Logger;
+}
 export interface CreateEntityServiceBaseParams extends EventParams {
   entityName: string;
   entity: Entity;
