+++ packages/amplication-code-gen-types/src/plugins-types.ts
@@ -57,6 +57,7 @@ export type PluginMap = {
 };
 
 export enum EventNames {
+  PrepareContext = "PrepareContext",
   CreateEntityController = "CreateEntityController",
   CreateEntityControllerBase = "CreateEntityControllerBase",
   CreateAuthModules = "createAuthModules",
